account for node configuration in graph editor
The objective of this commit is to correctly render graphs like these:
https://github.com/bhouston/behave-graph/blob/3e6568caa9a46a6f6aee27b0192a478f30c9c541/graphs/core/flow/WaitAll.json

When loading JSON, a node may have arbitrary configuration parameters,
that previously got ignored (such as number of sequence outputs).

This commit tracks configuration for each rendered node,
and accounts for it in all calculations (such as adding new edge).

In order to do this, instead of NodeSpecJSON (object) we now pass
a NodeSpecGenerator (class) which generates node specs on demand
based on passed configuration (and also memoizes, not sure if that's
necessary).

New nodes will still be generated with no configuration,
and no UI work to support it has been done.